requirements.txt
black == 23.12.1
jsonschema == 4.20.0
pre-commit == 3.6.0
pyyaml == 6.0.1
unidecode == 1.3.8
toml == 0.10.2
pandas == 2.2.3
openpyxl == 3.1.2
pyarrow == 12.0.0